museum scanner final/backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
pymongo
python-dotenv
